restart datasets directory focused on new provenance plans
