--- datasets/cache/filtered-industry.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:30b190027318116151ecf2b52adb5f3b8b3f12ba50f64f050b4fecf92b987b26
-size 393049931
